ONCOSCANAI/backend/requirements.txt

fastapi
uvicorn[standard]
python-multipart
Pillow
numpy
torch
torchvision
ultralytics
tensorflow
keras
